test(sandbox): address PR #44 review feedback (#52)
Extract duplicated gitInDir into shared testhelpers_test.go (was
copy-pasted in the linux and darwin integration_worktree files).
Extract the 4 copy-pasted run closures in grants_test.go into a
testGitRun helper. Simplify the dead inner guard in the submodule
test (subPath can never be under common). Narrow the omac read
grant from the whole temp dir to the binary itself. Document the
lenient gc assertion and the deliberately-broken repo state in
the symlink-escape test.
